# swagger.yaml
openapi: 3.0.0
info:
    title: Medical Booking API
    description: Made by - P's Get Degrees
    contact:
        name: Go to the Front End
        url: https://api-medical-booking.herokuapp.com/
    version: "1.0"
servers:
    - url: https://api-medical-booking.herokuapp.com/api/v1
      variables: {}
paths:
...
    /booking:
        get:
            tags:
                - booking
            summary: Get All Booking
            operationId: GetAllBooking
            parameters:
                - name: userId
                  in: query
                  schema:
                      type: string
                - name: doctorId
                  in: query
                  schema:
                      type: string
                - name: locationId
                  in: query
                  schema:
                      type: string
                - name: procedureId
                  in: query
                  schema:
                      type: string
                - name: week
                  in: query
                  description: "this is a timestamp that will indicate the range of the week and will return all bookings from sunday to saturday where the timestamp lies"
                  schema:
                      type: string
                      pattern: "^[0-9]+$"
                - name: dateFilters
                  in: query
                  required: false
                  description: "Possible variables ( bookingTime )  |  Supported operators ( > , >= , = , < , <= ) "
                  schema:
                      type: string
                      example: "bookingTime>=1633159914979"
                - name: select
                  in: query
                  required: false
                  description: "By default the query return all fields on the schema, defining a select field(s) will return on these field(s). Select filters can be chained using the ( , ) operator"
                  schema:
                      type: string
                      example: "doctorId,procedureId"
                - name: page
                  in: query
                  required: false
                  description: "By default the return data is segmetented depending on the limit of return and will return the first segment on the database. If the page value is 5 it will return the 5th segment of the database"
                  schema:
                      type: string
                      pattern: "^[0-9]+$"
                - name: limit
                  in: query
                  required: false
                  description: "By default the return data is 10 items. Setting the limit field to 100 will return 100 items."
                  schema:
                      type: string
                      pattern: "^[0-9]+$"
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
        post:
            tags:
                - booking
            summary: Create Booking
            operationId: CreateBooking
            parameters: []
            requestBody:
                content:
                    application/x-www-form-urlencoded:
                        encoding: {}
                        schema:
                            required:
                                - doctorId
                                - locationId
                                - procedureId
                                - description
                                - bookingTime
                            type: object
                            properties:
                                doctorId:
                                    type: string
                                locationId:
                                    type: string
                                procedureId:
                                    type: string
                                description:
                                    type: string
                                    description: Please describe your signs and symptoms
                                    example: This is a description
                                    minLength: 3
                                    maxLength: 100
                                bookingTime:
                                    type: string
                                    description: This booking time must be a timestamp
                                    example: 1632833484405
                required: false
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
    /booking/{id}:
        parameters:
            - in: path
              name: id
              schema:
                  type: string
              required: true
              description: The booking ID
        get:
            tags:
                - booking
            summary: Get Booking
            operationId: GetBooking
            parameters: []
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
        patch:
            tags:
                - booking
            summary: Update Booking
            operationId: UpdateBooking
            parameters: []
            requestBody:
                content:
                    application/x-www-form-urlencoded:
                        encoding: {}
                        schema:
                            type: object
                            properties:
                                doctorId:
                                    type: string
                                locationId:
                                    type: string
                                procedureId:
                                    type: string
                                description:
                                    type: string
                                    description: Please describe your signs and symptoms
                                    example: This is a description
                                    minLength: 3
                                    maxLength: 100
                required: false
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
        delete:
            tags:
                - booking
            summary: Delete Booking
            operationId: DeleteBooking
            parameters: []
            responses:
                "200":
                    description: ""
                    headers: {}
            deprecated: false
...
components:
    securitySchemes:
        httpBearer:
            type: http
            scheme: bearer
security:
    - httpBearer: []
...